requirements.txt
torch>=2.2
torchvision>=0.17
open-clip-torch>=2.24.0
Pillow
numpy
pandas
scikit-learn
opencv-python
streamlit
